ojus/scraper/requirements.txt
anthropic>=0.40.0
crawl4ai>=0.4.0
playwright>=1.40.0
httpx>=0.25.0
python-dotenv>=1.0.0
rich>=13.0.0
